Add support for parsing LaTeX in Markdown files
This implementation uses the UnifiedJS ecosystem to parse and process Markdown files located in posts/, with or without math blocks and LaTeX.

The processing takes place in lib/posts.ts, utilizing a handful of remark packages. Remark-math parses $ and $$ into math nodes. Remark-html-katex transforms those nodes with KaTeX. The necessary css for rendering KaTeX is imported in _app.tsx.